+++ CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Add `download` command ([#8])
 
 ## [0.2.0] - 2019-09-15
 ### Added
@@ -24,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 [0.2.0]: https://github.com/maxbrunet/mmemoji/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/maxbrunet/mmemoji/releases/tag/v0.1.0
 
+[#8]: https://github.com/maxbrunet/mmemoji/issues/8
 [#6]: https://github.com/maxbrunet/mmemoji/issues/6
 [#5]: https://github.com/maxbrunet/mmemoji/issues/5
 [#4]: https://github.com/maxbrunet/mmemoji/issues/4